+++ modin/core/io/text/fwf_dispatcher.py
@@ -18,8 +18,6 @@ from modin.core.io.text.text_file_dispatcher import TextFileDispatcher
 import pandas._libs.lib as lib
 import pandas
 from csv import QUOTE_NONE
-import sys
-import io
 
 from modin.config import NPartitions
 
@@ -48,204 +46,106 @@ class FWFDispatcher(TextFileDispatcher):
         new_query_compiler : BaseQueryCompiler
             Query compiler with imported data for further processing.
         """
-        filepath_or_buffer = cls.get_path_or_buffer(filepath_or_buffer)
-        if isinstance(filepath_or_buffer, str):
-            if not cls.file_exists(filepath_or_buffer):
-                return cls.single_worker_read(filepath_or_buffer, **kwargs)
-            filepath_or_buffer = cls.get_path(filepath_or_buffer)
-        elif not cls.pathlib_or_pypath(filepath_or_buffer):
-            return cls.single_worker_read(filepath_or_buffer, **kwargs)
-        compression_type = cls.infer_compression(
-            filepath_or_buffer, kwargs.get("compression", "infer")
+        filepath_or_buffer_md = (
+            cls.get_path(filepath_or_buffer)
+            if isinstance(filepath_or_buffer, str)
+            else cls.get_path_or_buffer(filepath_or_buffer)
         )
-        if compression_type is not None:
-            if (
-                compression_type == "gzip"
-                or compression_type == "bz2"
-                or compression_type == "xz"
-            ):
-                kwargs["compression"] = compression_type
-            elif (
-                compression_type == "zip"
-                and sys.version_info[0] == 3
-                and sys.version_info[1] >= 7
-            ):
-                # need python3.7 to .seek and .tell ZipExtFile
-                kwargs["compression"] = compression_type
-            else:
-                return cls.single_worker_read(filepath_or_buffer, **kwargs)
-
-        chunksize = kwargs.get("chunksize")
-        if chunksize is not None:
-            return cls.single_worker_read(filepath_or_buffer, **kwargs)
-
-        # If infer_nrows is a significant portion of the number of rows, pandas may be
-        # faster.
-        infer_nrows = kwargs.get("infer_nrows", 100)
-        if infer_nrows > 100:
-            return cls.single_worker_read(filepath_or_buffer, **kwargs)
-
-        skiprows = kwargs.get("skiprows")
-        if skiprows is not None and not isinstance(skiprows, int):
-            return cls.single_worker_read(filepath_or_buffer, **kwargs)
-        nrows = kwargs.pop("nrows", None)
+        compression_infered = cls.infer_compression(
+            filepath_or_buffer, kwargs.get("compression")
+        )
+        # Getting frequently used read_fwf kwargs
         names = kwargs.get("names", lib.no_default)
         index_col = kwargs.get("index_col", None)
-        if names in [lib.no_default, None]:
-            # For the sake of the empty df, we assume no `index_col` to get the correct
-            # column names before we build the index. Because we pass `names` in, this
-            # step has to happen without removing the `index_col` otherwise it will not
-            # be assigned correctly
-            names = pandas.read_fwf(
-                filepath_or_buffer,
-                **dict(kwargs, usecols=None, nrows=0, skipfooter=0, index_col=None),
-            ).columns
-        empty_pd_df = pandas.read_fwf(
-            filepath_or_buffer, **dict(kwargs, nrows=0, skipfooter=0)
+        encoding = kwargs.get("encoding", None)
+        skiprows = kwargs.get("skiprows", None)
+        header = kwargs.get("header", "infer")
+        # Define header size for further skipping (Header can be skipped because header
+        # information will be obtained further from empty_df, so no need to handle it
+        # by workers)
+        header_size = cls._define_header_size(
+            header,
+            names,
         )
-        column_names = empty_pd_df.columns
-        skipfooter = kwargs.get("skipfooter", None)
-        skiprows = kwargs.pop("skiprows", None)
-        usecols = kwargs.get("usecols", None)
-        usecols_md = cls._validate_usecols_arg(usecols)
-        if usecols is not None and usecols_md[1] != "integer":
-            del kwargs["usecols"]
-            with OpenFile(filepath_or_buffer, "rb") as f:
-                all_cols = pandas.read_fwf(
-                    f,
-                    **dict(kwargs, nrows=0, skipfooter=0),
-                ).columns
-            usecols = all_cols.get_indexer_for(list(usecols_md[0]))
-        parse_dates = kwargs.pop("parse_dates", False)
+        (
+            skiprows_md,
+            pre_reading,
+            skiprows_partitioning,
+        ) = cls._manage_skiprows_parameter(
+            skiprows, header_size
+        )  # TODO: add note about extending support of skiprows for read_fwf
+        should_handle_skiprows = skiprows_md is not None and not isinstance(
+            skiprows_md, int
+        )
+
+        use_modin_impl = cls._read_csv_check_support(
+            filepath_or_buffer, kwargs, compression_infered, check_fwf_specific=True
+        )
+        if not use_modin_impl:
+            return cls.single_worker_read(filepath_or_buffer, **kwargs)
+
+        is_quoting = kwargs.get("quoting", "") != QUOTE_NONE
+        quotechar = kwargs.get("quotechar", '"').encode(
+            encoding if encoding is not None else "UTF-8"
+        )
+        # In these cases we should pass additional metadata
+        # to the workers to match pandas output
+        pass_names = names in [None, lib.no_default] and (
+            skiprows is not None or kwargs.get("skipfooter", 0) != 0
+        )
+
+        pd_df_metadata = pandas.read_fwf(
+            filepath_or_buffer,
+            **dict(kwargs, nrows=1, skipfooter=0, index_col=index_col),
+        )
+        column_names = pd_df_metadata.columns
+        column_widths, num_splits = cls._define_metadata(pd_df_metadata, column_names)
+
+        # kwargs that will be passed to the workers
         partition_kwargs = dict(
             kwargs,
-            header=None,
-            names=names,
+            fname=filepath_or_buffer_md,
+            num_splits=num_splits,
+            header_size=header_size if not pass_names else 0,
+            names=names if not pass_names else column_names,
+            header=header if not pass_names else "infer",
             skipfooter=0,
             skiprows=None,
-            parse_dates=parse_dates,
-            usecols=usecols,
+            nrows=None,
+            compression=compression_infered,
         )
-        encoding = kwargs.get("encoding", None)
-        is_quoting = kwargs.get("quoting", "") != QUOTE_NONE
-        with OpenFile(filepath_or_buffer, "rb", compression_type) as f:
-            # Skip the header since we already have the header information and skip the
-            # rows we are told to skip.
-            if isinstance(skiprows, int) or skiprows is None:
-                if skiprows is None:
-                    skiprows = 0
-                header = kwargs.get("header", "infer")
-                if header == "infer" and kwargs.get("names", lib.no_default) in [
-                    lib.no_default,
-                    None,
-                ]:
-                    skiprows += 1
-                elif isinstance(header, int):
-                    skiprows += header + 1
-                elif hasattr(header, "__iter__") and not isinstance(header, str):
-                    skiprows += max(header) + 1
-            if kwargs.get("encoding", None) is not None:
-                partition_kwargs["skiprows"] = 1
-            # Launch tasks to read partitions
-            partition_ids = []
-            index_ids = []
-            dtypes_ids = []
-            column_widths, num_splits = cls._define_metadata(empty_pd_df, column_names)
 
-            args = {
-                "fname": filepath_or_buffer,
-                "num_splits": num_splits,
-                **partition_kwargs,
-            }
-
-            old_pos = f.tell()
-            fio = io.TextIOWrapper(f, encoding=encoding, newline="")
-            newline, quotechar = cls.compute_newline(
-                fio, encoding, kwargs.get("quotechar", '"')
-            )
-            f.seek(old_pos)
+        with cls.file_open(filepath_or_buffer_md, "rb", compression_infered) as f:
             splits = cls.partitioned_file(
                 f,
                 num_partitions=NPartitions.get(),
-                nrows=nrows,
-                skiprows=skiprows,
+                nrows=kwargs.get("nrows", None) if not should_handle_skiprows else None,
+                skiprows=skiprows_partitioning,
                 quotechar=quotechar,
                 is_quoting=is_quoting,
                 encoding=encoding,
                 newline=newline,
+                header_size=header_size,
+                pre_reading=pre_reading,
             )
-            for start, end in splits:
-                args.update({"start": start, "end": end})
-                partition_id = cls.deploy(cls.parse, num_splits + 2, args)
-                partition_ids.append(partition_id[:-2])
-                index_ids.append(partition_id[-2])
-                dtypes_ids.append(partition_id[-1])
 
-        # Compute the index based on a sum of the lengths of each partition (by default)
-        # or based on the column(s) that were requested.
-        if index_col is None:
-            row_lengths = cls.materialize(index_ids)
-            new_index = pandas.RangeIndex(sum(row_lengths))
-            # pandas has a really weird edge case here.
-            if (
-                kwargs.get("names", lib.no_default) not in [lib.no_default, None]
-                and skiprows > 1
-            ):
-                new_index = pandas.RangeIndex(
-                    skiprows - 1, new_index.stop + skiprows - 1
-                )
-        else:
-            index_objs = cls.materialize(index_ids)
-            row_lengths = [len(o) for o in index_objs]
-            new_index = index_objs[0].append(index_objs[1:])
-            new_index.name = empty_pd_df.index.name
-
-        # Compute dtypes by getting collecting and combining all of the partitions. The
-        # reported dtypes from differing rows can be different based on the inference in
-        # the limited data seen by each worker. We use pandas to compute the exact dtype
-        # over the whole column for each column. The index is set below.
-        dtypes = cls.get_dtypes(dtypes_ids)
-
-        partition_ids = cls.build_partition(partition_ids, row_lengths, column_widths)
-        # If parse_dates is present, the column names that we have might not be
-        # the same length as the returned column names. If we do need to modify
-        # the column names, we remove the old names from the column names and
-        # insert the new one at the front of the Index.
-        if parse_dates is not None:
-            # We have to recompute the column widths if `parse_dates` is set because
-            # we are not guaranteed to have the correct information regarding how many
-            # columns are on each partition.
-            column_widths = None
-            # Check if is list of lists
-            if isinstance(parse_dates, list) and isinstance(parse_dates[0], list):
-                for group in parse_dates:
-                    new_col_name = "_".join(group)
-                    column_names = column_names.drop(group).insert(0, new_col_name)
-            # Check if it is a dictionary
-            elif isinstance(parse_dates, dict):
-                for new_col_name, group in parse_dates.items():
-                    column_names = column_names.drop(group).insert(0, new_col_name)
-        # Set the index for the dtypes to the column names
-        if isinstance(dtypes, pandas.Series):
-            dtypes.index = column_names
-        else:
-            dtypes = pandas.Series(dtypes, index=column_names)
-        new_frame = cls.frame_cls(
-            partition_ids,
-            new_index,
-            column_names,
-            row_lengths,
-            column_widths,
-            dtypes=dtypes,
+        partition_ids, index_ids, dtypes_ids = cls._launch_tasks(
+            splits, **partition_kwargs
         )
-        new_query_compiler = cls.query_compiler_cls(new_frame)
 
-        if skipfooter:
-            new_query_compiler = new_query_compiler.drop(
-                new_query_compiler.index[-skipfooter:]
-            )
-        if kwargs.get("squeeze", False) and len(new_query_compiler.columns) == 1:
-            return new_query_compiler[new_query_compiler.columns[0]]
-        if index_col is None:
-            new_query_compiler._modin_frame.synchronize_labels(axis=0)
+        new_query_compiler = cls._get_new_qc(
+            partition_ids=partition_ids,
+            index_ids=index_ids,
+            dtypes_ids=dtypes_ids,
+            index_col=index_col,
+            index_name=pd_df_metadata.index.name,
+            column_widths=column_widths,
+            column_names=column_names,
+            skiprows_md=skiprows_md if should_handle_skiprows else None,
+            header_size=header_size,
+            squeeze=kwargs.get("squeeze", False),
+            skipfooter=kwargs.get("skipfooter", None),
+            parse_dates=kwargs.get("parse_dates", False),
+            nrows=kwargs.get("nrows", None) if should_handle_skiprows else None,
+        )
         return new_query_compiler
